+++ TeamEventApp/TeamEventApp/Classe/Event.cs
@@ -5,7 +5,7 @@ namespace TeamEventApp
 {
     public class Event
     {
-        public long eventId { get; set; }
+        public int eventId { get; set; }
         public string eventName { get; set; }
         public string description { get; set; }
         public DateTime startDate { get; set; }
